pyproject.toml
[tool.poetry]
name = "bloom"
version = "0.1.0"
description = "a compact library for discord."
authors = ["A5rocks <git@helvetica.moe>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
trio = "^0.19.0"
trio-websocket = "^0.9.2"
attrs = "^21.2.0"
cattrs = "^1.7.1"

typing-extensions = { version = "^3.10.0", python = "~3.7" }

[tool.poetry.dev-dependencies]
# docs
Sphinx = "^4.1.2"
furo = "^2021.7.28b40"
sphinx-autobuild = "^2021.3.14"
sphinxcontrib-trio = "^1.1.2"

# dev
mypy = "^0.910"
isort = "^5.9.2"
flake8 = "^3.9.2"
pylint = "^2.9.6"

# stubs
trio-typing = "^0.5.1"
typing-extensions = "^3.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
